+++ src/Service/Rekognition/src/RekognitionClient.php
@@ -27,6 +27,7 @@ use AsyncAws\Rekognition\Exception\ResourceNotFoundException;
 use AsyncAws\Rekognition\Exception\ResourceNotReadyException;
 use AsyncAws\Rekognition\Exception\ServiceQuotaExceededException;
 use AsyncAws\Rekognition\Exception\ThrottlingException;
+use AsyncAws\Rekognition\Input\CompareFacesRequest;
 use AsyncAws\Rekognition\Input\CreateCollectionRequest;
 use AsyncAws\Rekognition\Input\CreateProjectRequest;
 use AsyncAws\Rekognition\Input\DeleteCollectionRequest;
@@ -38,6 +39,7 @@ use AsyncAws\Rekognition\Input\IndexFacesRequest;
 use AsyncAws\Rekognition\Input\ListCollectionsRequest;
 use AsyncAws\Rekognition\Input\RecognizeCelebritiesRequest;
 use AsyncAws\Rekognition\Input\SearchFacesByImageRequest;
+use AsyncAws\Rekognition\Result\CompareFacesResponse;
 use AsyncAws\Rekognition\Result\CreateCollectionResponse;
 use AsyncAws\Rekognition\Result\CreateProjectResponse;
 use AsyncAws\Rekognition\Result\DeleteCollectionResponse;
@@ -54,6 +56,87 @@ use AsyncAws\Rekognition\ValueObject\Image;
 
 class RekognitionClient extends AbstractApi
 {
+    /**
+     * Compares a face in the *source* input image with each of the 100 largest faces detected in the *target* input image.
+     *
+     * If the source image contains multiple faces, the service detects the largest face and compares it with each face
+     * detected in the target image.
+     *
+     * > CompareFaces uses machine learning algorithms, which are probabilistic. A false negative is an incorrect prediction
+     * > that a face in the target image has a low similarity confidence score when compared to the face in the source
+     * > image. To reduce the probability of false negatives, we recommend that you compare the target image against
+     * > multiple source images. If you plan to use `CompareFaces` to make a decision that impacts an individual's rights,
+     * > privacy, or access to services, we recommend that you pass the result to a human for review and further validation
+     * > before taking action.
+     *
+     * You pass the input and target images either as base64-encoded image bytes or as references to images in an Amazon S3
+     * bucket. If you use the AWS CLI to call Amazon Rekognition operations, passing image bytes isn't supported. The image
+     * must be formatted as a PNG or JPEG file.
+     *
+     * In response, the operation returns an array of face matches ordered by similarity score in descending order. For each
+     * face match, the response provides a bounding box of the face, facial landmarks, pose details (pitch, roll, and yaw),
+     * quality (brightness and sharpness), and confidence value (indicating the level of confidence that the bounding box
+     * contains a face). The response also provides a similarity score, which indicates how closely the faces match.
+     *
+     * > By default, only faces with a similarity score of greater than or equal to 80% are returned in the response. You
+     * > can change this value by specifying the `SimilarityThreshold` parameter.
+     *
+     * `CompareFaces` also returns an array of faces that don't match the source image. For each face, it returns a bounding
+     * box, confidence value, landmarks, pose details, and quality. The response also returns information about the face in
+     * the source image, including the bounding box of the face and confidence value.
+     *
+     * The `QualityFilter` input parameter allows you to filter out detected faces that don’t meet a required quality bar.
+     * The quality bar is based on a variety of common use cases. Use `QualityFilter` to set the quality bar by specifying
+     * `LOW`, `MEDIUM`, or `HIGH`. If you do not want to filter detected faces, specify `NONE`. The default value is `NONE`.
+     *
+     * If the image doesn't contain Exif metadata, `CompareFaces` returns orientation information for the source and target
+     * images. Use these values to display the images with the correct image orientation.
+     *
+     * If no faces are detected in the source or target images, `CompareFaces` returns an `InvalidParameterException` error.
+     *
+     * > This is a stateless API operation. That is, data returned by this operation doesn't persist.
+     *
+     * For an example, see Comparing Faces in Images in the Amazon Rekognition Developer Guide.
+     *
+     * This operation requires permissions to perform the `rekognition:CompareFaces` action.
+     *
+     * @see https://docs.aws.amazon.com/rekognition/latest/dg/API_CompareFaces.html
+     * @see https://docs.aws.amazon.com/aws-sdk-php/v3/api/api-rekognition-2016-06-27.html#comparefaces
+     *
+     * @param array{
+     *   SourceImage: Image|array,
+     *   TargetImage: Image|array,
+     *   SimilarityThreshold?: float|null,
+     *   QualityFilter?: QualityFilter::*|null,
+     *   '@region'?: string|null,
+     * }|CompareFacesRequest $input
+     *
+     * @throws AccessDeniedException
+     * @throws ImageTooLargeException
+     * @throws InternalServerErrorException
+     * @throws InvalidImageFormatException
+     * @throws InvalidParameterException
+     * @throws InvalidS3ObjectException
+     * @throws ProvisionedThroughputExceededException
+     * @throws ThrottlingException
+     */
+    public function compareFaces($input): CompareFacesResponse
+    {
+        $input = CompareFacesRequest::create($input);
+        $response = $this->getResponse($input->request(), new RequestContext(['operation' => 'CompareFaces', 'region' => $input->getRegion(), 'exceptionMapping' => [
+            'AccessDeniedException' => AccessDeniedException::class,
+            'ImageTooLargeException' => ImageTooLargeException::class,
+            'InternalServerError' => InternalServerErrorException::class,
+            'InvalidImageFormatException' => InvalidImageFormatException::class,
+            'InvalidParameterException' => InvalidParameterException::class,
+            'InvalidS3ObjectException' => InvalidS3ObjectException::class,
+            'ProvisionedThroughputExceededException' => ProvisionedThroughputExceededException::class,
+            'ThrottlingException' => ThrottlingException::class,
+        ]]));
+
+        return new CompareFacesResponse($response);
+    }
+
     /**
      * Creates a collection in an AWS Region. You can add faces to the collection using the IndexFaces operation.
      *